Do not submit form when the user hits the `Enter` key when writing an answer
It's very discouraging for users to accidentally submit all the poll
answers form when hitting the Enter keyboard while writing an
open answer. This is the default behavior of text inputs, so
instead, we use a text area, which allows new lines, also making
it easier for users to introduce long answers to open questions.